tabs to send payment
